bevy_pbr
Function prepare_materials

pub fn prepare_materials<M: Material>(
    prepare_next_frame: Local<'_, PrepareNextFrameMaterials<M>>,
    extracted_assets: ResMut<'_, ExtractedMaterials<M>>,
    render_materials: ResMut<'_, RenderMaterials<M>>,
    render_device: Res<'_, RenderDevice>,
    images: Res<'_, RenderAssets<Image>>,
    fallback_image: Res<'_, FallbackImage>,
    pipeline: Res<'_, MaterialPipeline<M>>,
    default_opaque_render_method: Res<'_, DefaultOpaqueRendererMethod>,
)

This system prepares all assets of the corresponding Material type which where extracted this frame for the GPU.