bevy_pbr
Function prepare_preprocess_bind_groups

pub fn prepare_preprocess_bind_groups(
    commands: Commands<'_, '_>,
    render_device: Res<'_, RenderDevice>,
    batched_instance_buffers: Res<'_, BatchedInstanceBuffers<MeshUniform, MeshInputUniform>>,
    indirect_parameters_buffer: Res<'_, IndirectParametersBuffer>,
    mesh_culling_data_buffer: Res<'_, MeshCullingDataBuffer>,
    view_uniforms: Res<'_, ViewUniforms>,
    pipelines: Res<'_, PreprocessPipelines>,
)

A system that attaches the mesh uniform buffers to the bind groups for the variants of the mesh preprocessing compute shader.