bevy_pbr
Function queue_material_meshes

pub fn queue_material_meshes<M: Material>(
    (opaque_draw_functions, alpha_mask_draw_functions, transmissive_draw_functions, transparent_draw_functions): (Res<'_, DrawFunctions<Opaque3d>>, Res<'_, DrawFunctions<AlphaMask3d>>, Res<'_, DrawFunctions<Transmissive3d>>, Res<'_, DrawFunctions<Transparent3d>>),
    material_pipeline: Res<'_, MaterialPipeline<M>>,
    pipelines: ResMut<'_, SpecializedMeshPipelines<MaterialPipeline<M>>>,
    pipeline_cache: Res<'_, PipelineCache>,
    render_meshes: Res<'_, RenderAssets<RenderMesh>>,
    render_materials: Res<'_, RenderAssets<PreparedMaterial<M>>>,
    render_mesh_instances: Res<'_, RenderMeshInstances>,
    render_material_instances: Res<'_, RenderMaterialInstances<M>>,
    render_lightmaps: Res<'_, RenderLightmaps>,
    render_visibility_ranges: Res<'_, RenderVisibilityRanges>,
    opaque_render_phases: ResMut<'_, ViewBinnedRenderPhases<Opaque3d>>,
    alpha_mask_render_phases: ResMut<'_, ViewBinnedRenderPhases<AlphaMask3d>>,
    transmissive_render_phases: ResMut<'_, ViewSortedRenderPhases<Transmissive3d>>,
    transparent_render_phases: ResMut<'_, ViewSortedRenderPhases<Transparent3d>>,
    views: Query<'_, '_, (Entity, &ExtractedView, &RenderVisibleEntities, &Msaa, Option<&Tonemapping>, Option<&DebandDither>, Option<&ShadowFilteringMethod>, Has<ScreenSpaceAmbientOcclusion>, (Has<NormalPrepass>, Has<DepthPrepass>, Has<MotionVectorPrepass>, Has<DeferredPrepass>), Option<&Camera3d>, Has<TemporalJitter>, Option<&Projection>, (Has<RenderViewLightProbes<EnvironmentMapLight>>, Has<RenderViewLightProbes<IrradianceVolume>>), Has<OrderIndependentTransparencySettings>)>,
)
where
    M::Data: PartialEq + Eq + Hash + Clone,

For each view, iterates over all the meshes visible from that view and adds them to BinnedRenderPhases or SortedRenderPhases as appropriate.