Struct bevy_pbr::CascadeShadowConfigBuilder

pub struct CascadeShadowConfigBuilder {
    pub num_cascades: usize,
    pub minimum_distance: f32,
    pub maximum_distance: f32,
    pub first_cascade_far_bound: f32,
    pub overlap_proportion: f32,
}

Builder for CascadeShadowConfig.

Fields
num_cascades: usize

The number of shadow cascades. More cascades increases shadow quality by mitigating perspective aliasing - a phenomenon where areas nearer the camera are covered by fewer shadow map texels than areas further from the camera, causing blocky looking shadows.

This does come at the cost increased rendering overhead, however this overhead is still less than if you were to use fewer cascades and much larger shadow map textures to achieve the same quality level.

In case rendered geometry covers a relatively narrow and static depth relative to camera, it may make more sense to use fewer cascades and a higher resolution shadow map texture as perspective aliasing is not as much an issue. Be sure to adjust minimum_distance and maximum_distance appropriately.

minimum_distance: f32

The minimum shadow distance, which can help improve the texel resolution of the first cascade. Areas nearer to the camera than this will likely receive no shadows.

NOTE: Due to implementation details, this usually does not impact shadow quality as much as first_cascade_far_bound and maximum_distance. At many view frustum field-of-views, the texel resolution of the first cascade is dominated by the width / height of the view frustum plane at first_cascade_far_bound rather than the depth of the frustum from minimum_distance to first_cascade_far_bound.

maximum_distance: f32

The maximum shadow distance. Areas further from the camera than this will likely receive no shadows.

first_cascade_far_bound: f32

Sets the far bound of the first cascade, relative to the view origin. In-between cascades will be exponentially spaced relative to the maximum shadow distance. NOTE: This is ignored if there is only one cascade, the maximum distance takes precedence.

overlap_proportion: f32

Sets the overlap proportion between cascades. The overlap is used to make the transition from one cascade’s shadow map to the next less abrupt by blending between both shadow maps.

Implementations

impl CascadeShadowConfigBuilder

pub fn build(&self) -> CascadeShadowConfig

Returns the cascade config as specified by this builder.