Struct bevy_pbr::StandardMaterialUniform

pub struct StandardMaterialUniform {

    pub base_color: Vec4,
    pub emissive: Vec4,
    pub attenuation_color: Vec4,
    pub uv_transform: Mat3,
    pub roughness: f32,
    pub metallic: f32,
    pub reflectance: f32,
    pub diffuse_transmission: f32,
    pub specular_transmission: f32,
    pub thickness: f32,
    pub ior: f32,
    pub attenuation_distance: f32,
    pub clearcoat: f32,
    pub clearcoat_perceptual_roughness: f32,
    pub anisotropy_strength: f32,
    pub anisotropy_rotation: Vec2,
    pub flags: u32,
    pub alpha_cutoff: f32,
    pub parallax_depth_scale: f32,
    pub max_parallax_layer_count: f32,
    pub lightmap_exposure: f32,
    pub max_relief_mapping_search_steps: u32,
    pub deferred_lighting_pass_id: u32,

}

The GPU representation of the uniform data of a StandardMaterial.

Fields
base_color: Vec4

Doubles as diffuse albedo for non-metallic, specular for metallic and a mix for everything in between.

emissive: Vec4attenuation_color: Vec4

Color white light takes after travelling through the attenuation distance underneath the material surface

uv_transform: Mat3

The transform applied to the UVs corresponding to ATTRIBUTE_UV_0 on the mesh before sampling. Default is identity.

roughness: f32

Linear perceptual roughness, clamped to [0.089, 1.0] in the shader Defaults to minimum of 0.089

metallic: f32

From [0.0, 1.0], dielectric to pure metallic

reflectance: f32

Specular intensity for non-metals on a linear scale of [0.0, 1.0] defaults to 0.5 which is mapped to 4% reflectance in the shader

diffuse_transmission: f32

Amount of diffuse light transmitted through the material

specular_transmission: f32

Amount of specular light transmitted through the material

thickness: f32

Thickness of the volume underneath the material surface

ior: f32

Index of Refraction

attenuation_distance: f32

How far light travels through the volume underneath the material surface before being absorbed

clearcoat: f32clearcoat_perceptual_roughness: f32anisotropy_strength: f32anisotropy_rotation: Vec2flags: u32

The StandardMaterialFlags accessible in the wgsl shader.

alpha_cutoff: f32

When the alpha mode mask flag is set, any base color alpha above this cutoff means fully opaque, and any below means fully transparent.

parallax_depth_scale: f32

The depth of the StandardMaterial::depth_map to apply.

max_parallax_layer_count: f32

In how many layers to split the depth maps for Steep parallax mapping.

If your parallax_depth_scale is >0.1 and you are seeing jaggy edges, increase this value. However, this incurs a performance cost.

lightmap_exposure: f32

The exposure (brightness) level of the lightmap, if present.

max_relief_mapping_search_steps: u32

Using ParallaxMappingMethod::Relief, how many additional steps to use at most to find the depth value.

deferred_lighting_pass_id: u32

ID for specifying which deferred lighting pass should be used for rendering this material, if any.