bevy_pbr
Struct ViewLightProbesUniformOffset

pub struct ViewLightProbesUniformOffset(/* private fields */);

A component attached to each camera in the render world that stores the index of the LightProbesUniform in the LightProbesBuffer.