bevy_pbr
Type Alias VolumetricFogSettings

pub type VolumetricFogSettings = VolumetricFog;

👎Deprecated since 0.15.0: Renamed to VolumetricFog

Aliased Type

struct VolumetricFogSettings {
    pub ambient_color: Color,
    pub ambient_intensity: f32,
    pub jitter: f32,
    pub step_count: u32,
}

Fields
ambient_color: Color

Color of the ambient light.

This is separate from Bevy’s AmbientLight because an EnvironmentMapLight is still considered an ambient light for the purposes of volumetric fog. If you’re using a EnvironmentMapLight, for best results, this should be a good approximation of the average color of the environment map.

Defaults to white.

ambient_intensity: f32

The brightness of the ambient light.

If there’s no EnvironmentMapLight, set this to 0.

Defaults to 0.1.

jitter: f32

The maximum distance to offset the ray origin randomly by, in meters.

This is intended for use with temporal antialiasing. It helps fog look less blocky by varying the start position of the ray, using interleaved gradient noise.

step_count: u32

The number of raymarching steps to perform.

Higher values produce higher-quality results with less banding, but reduce performance.

The default value is 64.