bevy_pbr
Module wireframe

Structs

NoWireframe

Disables wireframe rendering for any entity it is attached to. It will ignore the WireframeConfig global setting.

Wireframe

Enables wireframe rendering for any entity it is attached to. It will ignore the WireframeConfig global setting.

WireframeColor

Sets the color of the Wireframe of the entity it is attached to.

WireframeConfig

WireframeMaterial

WireframePlugin

A Plugin that draws wireframes.

Constants

WIREFRAME_SHADER_HANDLE