Module bevy_pbr::deferred

Structs

DeferredLightingLayout

DeferredLightingPipeline

DeferredOpaquePass3dPbrLightingNode

DeferredPbrLightingPlugin

PbrDeferredLightingDepthId

Component with a depth_id for specifying which corresponding materials should be rendered by this specific PBR deferred lighting pass. Will be automatically added to entities with the DeferredPrepass component that don’t already have a PbrDeferredLightingDepthId.

Constants

DEFAULT_PBR_DEFERRED_LIGHTING_PASS_ID

DEFERRED_LIGHTING_SHADER_HANDLE

Functions

insert_deferred_lighting_pass_id_component

prepare_deferred_lighting_pipelines